+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 8.0.3
+* Added a "Start Destroying" button to the Dejunk frame
+* Set "Verbose Mode" to be enabled by default
+* Fixed a bug causing "Below Average Equipment" to ignore poor quality items
+* Updated libraries
+* Various code improvements
+
+*****
+
 ### 8.0.2
 * Updated ruRU (Russian) localization